Role based policy implementation
It adds role bases access control in all openstack apis of
MagnetoDB.The policy file lives in /etc/magnetodb/policy.json.

A rule "mdb:create_table":"role:admin or tenant:(tenant_id)s"
means that anyone having role admin or anyone having a token
scoped to project whose project_id is present in the url can
create a table.

A rule "mdb:list_tables":"!" means that nobody would be allowed
to access the api applying this check.

A rule "mdb:list_tables":"" means that everyone is allowed to
access the api.

The rule "default" applies for the rules which are not present
in the policy.json file. For example if you remove all the rules
and keep just "default":"" then all the apis are accessible by
everyone.

bp : support-roles

Change-Id: I213aea56c47261d90a2ab0f8b5042bae9da3d8bc