Change payload in notification to contain uuid
The payload is changed to contain table_uuid
along with table_name. Cases where uuid is
not available, the field would be an empty string.
A new field called index.count is also added when
magnetodb.table.create.end is genarated. This is
generated to keep a count of indices in Ceilometer.
Also updated the log messages to contain uuid.

Changed the user field in context to contain
user_id instead of user_name.

Closes-bug: #1400695

Change-Id: Ia6bfbf929945f922382dd7d6a0965863edd5b8b7