Refactor notification
A middleware is introduced to send all API request level notifications using
a refactored notifier. For asynchronous requests such as table create/delete,
explicit notifications will be used to send events.

A new notifier is introduced to send events and/or other metrics(to be implemented),
depending on the event type defined in a central event registry. Notification
is refactored to use the new notifier.

Change-Id: I7c55e4385fd8337256ee6a7b6b884dfbbebb0a2c
Implements: blueprint refactor-notification